Moving Nvidia interfaces to stats package so that they can be used outside of the accelerators package
