+++ stats/types.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Google Inc. All Rights Reserved.
+// Copyright 2020 Google Inc. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,22 +11,25 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package accelerators
+
+// Handling statistics that are fully controlled in cAdvisor
+package stats
 
 import info "github.com/google/cadvisor/info/v1"
 
-// This is supposed to store global state about an accelerator metrics collector.
-// cadvisor manager will call Setup() when it starts and Destroy() when it stops.
-// For each container detected by the cadvisor manager, it will call
+// This is supposed to store global state about an cAdvisor metrics collector.
+// cAdvisor manager will call Setup() when it starts and Destroy() when it stops.
+// For each container detected by the cAdvisor manager, it will call
 // GetCollector() with the devices cgroup path for that container.
 // GetCollector() is supposed to return an object that can update
 // accelerator stats for that container.
-type AcceleratorManager interface {
+type Manager interface {
 	Setup()
 	Destroy()
-	GetCollector(deviceCgroup string) (AcceleratorCollector, error)
+	GetCollector(deviceCgroup string) (Collector, error)
 }
 
-type AcceleratorCollector interface {
+// Collector can update ContainerStats by adding more metrics.
+type Collector interface {
 	UpdateStats(*info.ContainerStats) error
 }
